requirements.txt
streamlit
requests
python-dotenv
firebase-admin
streamlit-authenticator
streamlit-option-menu
streamlit-elements
